+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Next Release
 
+## 1.24.7
+
 - Fix a bug where clicking "+Vow" would create a "progress" item instead of a "vow" item.
 - Extend the location sheet to include IS islands and settlements ([#1037](https://github.com/ben/foundry-ironsworn/pull/1037))